% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episodes.R
\name{mod_episodes_ui}
\alias{mod_episodes_ui}
\alias{mod_episodes_server}
\title{mod_episodes_ui and mod_episodes_server}
\usage{
mod_episodes_ui(id)

mod_episodes_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
