% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modeling_budget.R
\name{mod_modeling_budget_ui}
\alias{mod_modeling_budget_ui}
\alias{mod_modeling_budget_server}
\title{mod_modeling_budget_ui and mod_modeling_budget_server}
\usage{
mod_modeling_budget_ui(id)

mod_modeling_budget_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
