% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modeling_scenario.R
\name{mod_modeling_scenario_ui}
\alias{mod_modeling_scenario_ui}
\alias{mod_modeling_scenario_server}
\title{mod_modeling_scenario_ui and mod_modeling_scenario_server}
\usage{
mod_modeling_scenario_ui(id)

mod_modeling_scenario_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
