% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navbar.R
\name{mod_navbar_ui}
\alias{mod_navbar_ui}
\alias{mod_navbar_server}
\title{mod_navbar_ui and mod_navbar_server}
\usage{
mod_navbar_ui(id)

mod_navbar_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
