% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_score.R
\name{mod_score_ui}
\alias{mod_score_ui}
\alias{mod_score_server}
\title{mod_score_ui and mod_score_server}
\usage{
mod_score_ui(id)

mod_score_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
